add search track test
